Use object-fit:contain to maintain thumbnail aspect ratio
Currently, thumbnails that aren't the same aspect ratio as the <img> will have their ratio warped to fit the img element. Object-fit:contain will stretch the object to fit the img box, but maintain the aspect ratio.

Additionally, the aspect ratio of the img boxes currently aren't 3:2 because of the different "top/bottom" and "left/right" values. Compensate for this in by adding "* 1.05" padding-top in parent element.